Many, many changes:
Using COLORNWT as node weight type for MWIS calls.
It can be found in the new header color_defs.h

Added a heap and a new greedy routine that uses the
node weight minus the sum of the weights of free neighbors.

The MWIS greedy code will try up to 3 strategies now,
if the preceding failed.