New feature for multi-processor version:
The code will now keep track of the age a job and give a note once
a job grows very old. This is an indicator that a worker has died.
The new program color_jobkiller can be used to remove that job
from the list and re-insterted to the heap.